+++ src/flatpickr.directive.ts
@@ -391,7 +391,7 @@ export class FlatpickrDirective
 
   registerOnTouched(fn: any): void {} // tslint:disable-line
 
-  @HostListener('change')
+  @HostListener('input')
   inputChanged(): void {
     const value: string = this.elm.nativeElement.value;
     if (this.convertModelValue && typeof value === 'string') {
@@ -410,6 +410,8 @@ export class FlatpickrDirective
         default:
           this.onChangeFn(new Date(value));
       }
+    } else {
+      this.onChangeFn(value);
     }
   }
 }